fix race around span list by protecting its access with a lock (#7)
If a span were to arrive right around when a trace's timer went off, you'd get concurrent access to the list of spans when the collector tries to add a new span while the sending function is trying to iterate across those spans. This can cause the iterator to get a nil span out of the list.

In addition to protecting the span list with a lock this change removes the nil span protection that was in the dynsampler because it should no longer be necessary; there is no case where it should get a nil span. A panic there is nice to help make sure that's true.