add a larger sent trace cache (#11)
When we overrun the trace cache, terrible things happen.
* the trace gets orphaned so no new spans (notably the root span) can not be added to it
* the orphaned trace will always wait the full 60sec delay before sending
* we lose the record of how we chose to sample the trace so new spans coming in to the trace will make a new trace and get a different decision 

This change separates the "did I send this trace" cache from the "let me hold this trace structure to build up a tree and make a sampling decision" cache. It creates a much larger (5x) sent cache, since the only thing that needs to exist in the sent cache is a record of whether the trace was sent to Honeycomb and if so, the sample rate used. 

By separating these two caches, we can then also change the regular trace accumulating cache to immediately send traces that are getting evicted, since the record of how they were sampled can stick around. In this new world, future spans that come in to a trace that got evicted from the cache can still honor the same sampling decision and get immediately sent along their way.

This should really calm down the explosion in memory used when we overrun the trace cache, since the whole trace object won't be hanging around orphaned (but still referenced by an outstanding goroutine).