read response queues, print errored responses (#26)
It'd be nice to know when the place we're sending things hands back errors, whether the things being sent be events, spans, logs, or metrics, and whether we're sending them to peers or upstream to Honeycomb itself. This change teaches samproxy to read the responses queue and do things with errors it gets back down that queue.