put a buffered channel in between accepting spans and navigating the internal data structures to avoid blocking the client (#28)
There are a bunch of locks involved in getting a span into the right in-memory data structure representing a trace. Previously, the client sending us a span or a batch of spans must wait for all the internal bookkeeping to finish before it gets its response. This can be slow and, especially when dealing with a peer to peer transmission, cause thundering herd backups. Separating receipt of the span with the internal bookkeeping via a buffered channel will allow the recipient to spread out the load with the channel accepting the expected ebb and flow of traffic.