single thread the collector, remove all the collector's locks (#32)
When used with the dynamic sampler and heavy load, the previous version of samproxy exhibited a thundering herd kind of behavior where a bunch of spans would get sent and then everything would lock up for a while. I believe one cause is all the locks that were scattered through the collector, the collector's cache, and the trace object.

This change removes all those locks by single threading the collector and using channels to serialize all activity. The collector is the thing that

    inserts spans into traces
    calls out to the dynamic sampler to make a sampling decision
    hands off spans to the transmission to be sent

The timeouts (the biggest reason for concurrent access to all the trace objects) are now handled by a goroutine that waits and then submits the trace for sending by adding it to another channel. The main collect entry point reads from the incoming span channel and the ready-to-be-sent trace channel, dealing with them as they come in.

Single threading the collector removed 5 locks. Curious to try it out and see what happens.