preferentially process peer traffic to allow upstream backpressure without interal traffic jams (#33)
When samproxy is fed high volumes of traffic, it has the potential to get itself locked up. Traffic is accepted from clients and routed to peers. Accepting more traffic from clients gets blocked because a samproxy host is waiting for its peer to accept spans. But the peer is itself blocked waiting for other peers to accept traffic. This leads to a gridlock-style traffic jam.

By preferentially processing spans coming from peers over spans coming from the thing actually sending Samproxy data, we'll ensure that the peer's traffic is processed first, and that traffic from clients is only added in to the system when there's enough room to deal with all the peer traffic. This will exert backpressure on clients while keeping the roads clear for p2p traffic to flow.