namespace metrics for transmission to better understand behavior (#36)
Prometheus prohibits you from registering the same metric multiple times. Understandably, as having multiple sources try and own the same metric name makes it difficult to get a good understanding of what has happened when the metric is incremented.

When peer traffic was separated from incoming traffic, the Transmission object got registered twice; once for each type of traffic. Unfortunately, it used the same metric names for both, causing the process to panic on start when configured with a prometheus exporter.

This change namespaces the metrics used in the Transmission to avoid this conflict.