add option to use redis for managing peer membership in the cluster (#35)
Configuring the list of peers in a static configuration file is annoying in the face of clusters that may have nodes fail and be replaced. Far better is to use a shared configuration system so that peers can register themselves and get a list of all live peers at any time.

When a peer stops, it should fall out of the configuration and the cluster should continue to function with one fewer member. When new nodes come online, they should automatically join the cluster.

This change adds this behavior using Redis as the shared storage for the list of peers participating in the cluster.