sharder: Refuse to reload empty peer list
The peer list can temporarily become empty when using Redis discovery
and the Redis state is lost (e.g. restarting an in-memory instance).
This causes a divide by zero panic when attempting to chose a shard and
results in dropping spans that may have already been correctly routed by
another peer.

I experienced this while running a load test to confirm the behaviour
when Redis becomes unavailable. It continues working fine while Redis is
unavailable, but it resulted in these panics when Redis came back:

    $ kc logs samproxy-96765ff85-44qg5
    time="2020-06-10T14:13:34Z" level=info msg="using identifier from interface" identifier=172.17.0.7 interface=eth0
    time="2020-06-10T14:15:52Z" level=error msg="redis scan encountered an error" err=EOF keys_returned=0 timeoutSec=5s
    time="2020-06-10T14:15:52Z" level=error msg="redis scan encountered an error" err=EOF keys_returned=0 timeoutSec=5s
    time="2020-06-10T14:15:52Z" level=error msg="handler returning error" error.err="caught panic: runtime error: integer divide by zero" error.msg="caught panic" error.status_code=500
    time="2020-06-10T14:15:53Z" level=error msg="handler returning error" error.err="caught panic: runtime error: integer divide by zero" error.msg="caught panic" error.status_code=500
    time="2020-06-10T14:15:53Z" level=error msg="handler returning error" error.err="caught panic: runtime error: integer divide by zero" error.msg="caught panic" error.status_code=500
    $ kc logs samproxy-96765ff85-2zkzw
    time="2020-06-10T14:13:32Z" level=info msg="using identifier from interface" identifier=172.17.0.10 interface=eth0
    time="2020-06-10T14:15:52Z" level=error msg="non-20x response when sending event" api_host="http://172.17.0.7:8081" dataset=beeline-go event_type=unknown status_code=500 target=unknown
    time="2020-06-10T14:15:52Z" level=error msg="non-20x response when sending event" api_host="http://172.17.0.7:8081" dataset=beeline-go event_type=unknown status_code=500 target=unknown
    time="2020-06-10T14:15:52Z" level=error msg="non-20x response when sending event" api_host="http://172.17.0.7:8081" dataset=beeline-go event_type=unknown status_code=500 target=unknown

No spans were dropped after repeating the test with this change. There's
always going to be a possible race condition with some inconsistent
sharding/sampling as the peers re-register and refresh the lists at
different intervals, but this will deal with the worst case.

Returning an error when there are no peers means that it will surface
the error to the logs and re-attempt later. I considered putting the
check beneath the URL parsing but there's no benefit since that will
return an error rather than continuing without a peer.

We originally discussed only keeping the peer list if there was
previously a non-zero peer list but that seems unnecessary to check
since it can't start with a non-zero peer list (`Start` will return an
error if it can't find itself).