collect: Fix some missing metrics
These metrics weren't being reported because they were registered with
the wrong name or not registered at all. I originally noticed that
duration and span count were always zero in our installation and then
found the others with:

    $ for metric in $(rg '.*(IncrementCounter|Histogram)\("(.*)".*' -Ir '$2'); do rg -q "Register\(\"$metric\"" || echo $metric; done
    collector_tosend_queue
    trace_span_count
    trace_duration_ms

The `metrics` package could be made more resilient or more obvious to
this happening in the future.