split main and rules config files (#77)
The current config file is a behemoth and it’s not clear which parts are required, which are for tuning, and which are for sampling. It’s not clear which you set once at the beginning and which you would manage more frequently as your traffic changes. This is the first step in moving towards a better world by splitting the sampling rules config from the main config file.