fix broken log line, add WithString for efficiently skipped logs (#133)
Turns out logrus doesn't do what I want when I pass a pointer to a string to WithField. Instead, add WithString() to the Logger interface, which avoids the heap allocation for skipped logs. It's a bit awkward to have both but I think that's a small price to pay for a simple way to avoid vast quantities of garbage.

For background, the problem with WithField() is that passing a scalar into the interface{} value param causes it to escape to the heap, even if it's ultimately not used. This can add up to a lot of allocations for frequent log lines, and is a huge waste when those logs are disabled. However, a typed string value doesn't escape here, so calling WithString() is essentially free for disabled log levels.