add distributed benchmark (#135)
Adds a new benchmark with a 5-node cluster. Results are pretty much as you'd expect, but profiling this does show that we spend a lot of time locking metrics. Maybe not a chief concern at this time, but something to keep an eye on.

Also includes a shutdown code for Router.