implement optional soft allocation ceiling (#139)
Adds a new MaxAlloc config value. If > 0, InMemCollector will periodically check to see if the process's total allocation is above the limit. If it is, it will send and discard 10% of traces, and reduce the cache capacity to the new size. For this to work well, it should be set to about 75%-90% of available RAM, to leave room for scratch space and various fluctuations.

The intention is that this is a failsafe for if the cache capacity has been set too high; if it triggers regularly, users should lower their starting cache capacity and/or add machines to their samproxy cluster.