add mutex locking to fileConfig (#143)
The trouble with config reloads is that it means we've got a background goroutine writing new config values - at the same time our code is occasionally reading these values. The real-world risk here is pretty low but in the spirit of cleaning up go -race it seems worth getting out of the way.