add blank rules file to build-pkg, use old rules file as example (#196)
In adding more detail to the getting started docs, we noticed that the rules file had been left out of the build package. Since the existing rules file was more of an example file, and so as not to pollute the config with those examples, we renamed rules.toml to rules_complete.toml. Since defaults for the rules configuration are set in file_config.go, rules.toml is now blank until folks start adding their own rules.

Drive-by cleanup of a few things in file_config.go and regarding our config tests.