Make gRPC ServerParameters configurable
Which problem is this PR solving
--------------------------------

At the moment, there is no way to override the ServerParameters used in
the `keepalive` configuration of the gRPC server. As to @bdarfler's
comments in #263, issues can arise where new hosts will never be
connected to because clients will reconnect to already existing hosts
and by the time replacement hosts are available, all clients will have
reconnected to existing hosts with a
`MaxConnectionAge`/`MaxConnectionAgeGrace` of
[infinity](https://github.com/grpc/grpc-go/blob/60a3a7e969c401ca16dbcd0108ad544fb35aa61c/internal/transport/defaults.go#L36-L37).

Short description of the changes
--------------------------------

The changes made are to plumb through the configuration of the GRPC
ServerParameters from the viper configuration and into the instantiation
of the actual gRPC server. These changes kept the previous default
configuration of refinery as the defaults of the configurable settings,
so anyone who doesn't explicitly override any of these configurations
should have exactly the same behavior.

Closes #263