Extract Sent Cache to an interface for future expansion (#561)
## Which problem is this PR solving?

This is a prep PR for further work on the sent trace cache. For improved
scalability, we want to be able to store trace decision records for a
longer time. The best way to implement this in a backwards-compatible
way is to pull the mechanisms for managing decision records into an
interface, and then implement the interface with the legacy logic.
That's what this does. There are no expected changes in behavior, and
all tests still pass.

## Short description of the changes

- Define interfaces for a TraceSentCache and a TraceSentRecord
- Implement code for those that duplicates the existing legacy logic
- Refactor the places the code is used to use the new interfaces
- Tweak span count data type so that it's not an int64 any more
- Rename SpanCount to DescendantCount because now that we have links and
events, they're not all Spans anymore, and future PRs will add
additional counting functions. I haven't renamed the corresponding Get
and Add functions because that's pretty messy.