feat: Add ability to set Redis database and prefix in config (#614)
## Which problem is this PR solving?

In a blue/green deployment model, two refinery clusters might want to
share the same instance of redis. It's also possible that multiple apps
want to share one redis. This permits that by:
- allowing the redis "database" (an integer from 0-15) to be specified
in the configuration
- allowing the redis "prefix" (a string used in the name of the keys
stored in redis) to be set as well. It already existed but was hardcoded
to "refinery".

## Short description of the changes

- Add RedisDatabase and RedisPrefix to the PeerManagement section of
config.
- Add tests
- Add documentation

Fixes #544