feat: Refactoring of rules configuration and rules converter (#681)
## Which problem is this PR solving?

This is a reworking of the refinery rules structure. The existing rules
configuration format is both awkwardly structured for technical reasons,
as well as just being confusing. People make a lot of mistakes.

This structure organizes refinery rules into a map of environment name
(dataset in classic mode) to sampler config, and reserves the name
`__default__` for the default case. It also puts the sampler
configurations indented under the sampler type, which avoids another
level of confusion.

## Short description of the changes

- Refactor the sampler_config file so all the configuration types are
near each other
- Rename the Rule and Condition slices to Rules and Conditions
- Decorate all the types with yaml (and sometimes json) tags
- Add a "rules" command to the conversion tool that uses this stuff to
do the conversion

This work isn't "done" (there is more planned) but it's stable enough to
call it a checkpoint, review as-is, and do the next things as separate
PRs.