feat: Remove trace key params -> add meta sample_key (#685)
## Which problem is this PR solving?

Refinery had rules configuration for dynamic samplers to
"AddSampleRateKeyToTrace" which would inject the sample rate key into a
field of the user's choice. We also have, in the normal configuration
"AddRuleReasonToTrace", which also injected that key but as part of the
rule reason in `meta.refinery.reason`, causing people to have to write
derived columns to pull it off.

This PR addresses both of those.

## Short description of the changes

- Remove `AddSampleRateKeyToTrace` and `AddSampleRateKeyToTraceField`
params from samplers
- Remove all the tests for it
- Modify the GetSampleRate function to return the reason and key
separately (or a blank key)
- If the key isn't blank, add it to the spans under
`meta.refinery.sample_key`
- Add tests for it

This looks big but it's a lot of small identical changes in multiple
files.