fix: send probes to peers during stress relief (#689)
## Which problem is this PR solving?

- #688 

When stress relief activates, it stops sending peer traffic. But the
metrics that update stress relief based on peer volume only update when
there's peer traffic.

This causes any trace that refinery keeps (forwards to Honeycomb) in
stress relief mode to *also* be sent as a probe to the appropriate peer
(which then drops it). This ensures that a little bit of traffic also
gets to peers, which keeps the system operating normally and updating
metrics.

## Short description of the changes

- Make a copy of the incoming event as a probe whenever stress relief is
active and would keep a span; if the probe would have been sent to a
peer, send it.
- When a probe is received from a peer, drop it without further
examination.
- Log when this happens on both sides.
- Add the stress relief reason to the `meta.refinery.reason` field for
kept spans so that it's not only found in the logs.

---------

Co-authored-by: Alyson van Hardenberg <avanhardenberg@honeycomb.io>