feat: [v2] Use the new configuration system (#690)
This converts refinery to use the new configuration model that we've
been moving towards

This is an enormous PR because there really wasn't a good way to change
to a new configuration model slowly; it affected a lot of support code
and tests.

Key elements include:
* The main configuration was reorganized into a collection of structs,
each of which has appropriate struct tags for validation and defaults.
* All the configuration support code now uses the new configuration
structs. Some of the function names have been changed, and more will be
changed in this or later PRs.
* The Config interface changed to support this, which means that the
mock config also changed.
* Rules have been restructured so that the rules file can be read
without pre-reading it to understand its structure. This also enables it
to be validated normally.
* All duration-related values are now specified as Duration objects and
must be specified with units (like `3m30s` or `100ms`).
* Code supporting the legacy sharding system was removed. This required
updating a couple of tests that had been designed around deterministic
traceIDs.
* SampleCache no longer supports the legacy trace caching mechanism. Its
code was deleted and tests updated.
* The "resize" cache management strategy was removed. Its code was
deleted and tests updated.
* Tests requiring configuration were changed to generate YAML from a map
rather than trying to embed YAML into the test scripts. This shortened
them a lot and makes them easier to maintain.
* Along the way, several issues were uncovered in the configuration as
it had been redefined and so those items were updated and the template
regenerated.

There is still work to do before this can be released, but it's passing
tests and I can go back to smaller PRs after this.

This closes a few of the planned issues for 2.0:
Closes #667 
Closes #655 
Closes #581