feat: validation integration, part 1 (#700)
## Which problem is this PR solving?

This fleshes out the validation config file and makes it really work.

The validation system needed to be part of Refinery itself; I tried
putting it in its own library, but it made more sense to make it an
integral part of config.

## Short description of the changes

- Moves the validation logic into Config
- Rename the validation data file and instances of it to metadata and
into a validation object
- Use the object in the convert tool
- Embed the metadata file into the binary
- Lots of tests of the validation logic
- Update the metadata file to include all the validations
- It successfully checks the complete config file without errors

This does not do rule validation yet -- that's next.

This also does not integrate the validation into Refinery's loading yet,
although that is just a few lines of code. That will come after rule
validation.