feat: write out parsed configs (#707)
## Which problem is this PR solving?

We have often been asked for the ability to see what refinery thinks its
configurations are. We support this with a debug command, but it
requires a running refinery configured to allow debug queries, and a
configuration that allows connecting to it.

This PR adds command-line switches to read the configs, apply defaults
and environment variables, and write the data back out to a file. If you
use this feature, refinery doesn't start, it just writes the file and
exits.

## Short description of the changes

- Add function to write YAML out
- Add cmdline parameters to control it

Dependent on #706