fix: Update dynamic samplers to use GetSampleRateMulti (#717)
## Which problem is this PR solving?

- As it says in #636, throughput and dynamic sampler calculations have
been based on the wrong numbers for some time. They've been counting
traces, not spans, and yet everything about Honeycomb is based on span
count. Most people think it works by counting spans and are then
surprised when sample rates seem incorrect for their volume, and in
fact, our documentation says that it works this way.

## Short description of the changes

- This changes Refinery's dynamic and throughput sampler requests to use
the span count, not the trace count, when calculating numbers for the
dynamic samplers and throughput samplers.

This will need to be reflected in updated documentation.

Closes #636.