feat: Warn about samplers that might need adjustment (#718)
## Which problem is this PR solving?

- Since we adjusted sampler behavior to count spans, not traces, users
may need to tweak sample rates and throughput values. This change has
the converter issue a warning in those cases.
- Resolves a concern @TylerHelmuth expressed in #717 

## Short description of the changes

- Add test for which samplers might need to be fixed
- Check it on rule conversion

The warning looks like this for the demo rules file:
```
WARNING: Version 2 of Refinery has changed the way that sample rates are calculated for
the dynamic and throughput samplers. Refinery v1.x was documented as counting the number
of spans, but in fact it was counting traces. Version 2 samplers correctly count the
number of spans when doing these calculations.

This means that you may need to adjust the target throughput or sample rates to get the
same behavior as before.

When using v1, if you have had difficulty reaching your target throughput or sample rates,
or if different parts of your key space vary widely in their span counts, then there is a
good chance you should lower the target values in v2.

The following parts of your configuration should be examined:
 *  __default__: DeterministicSampler
 *  dataset1: DynamicSampler
 *  dataset2: EMADynamicSampler
 *  dataset3: DeterministicSampler
 *  dataset4: RulesBasedSampler(dynamically sample 200 responses, downstream Sampler EMADynamicSampler)
 *  dataset4: RulesBasedSampler(dynamically sample 200 string responses, downstream Sampler EMADynamicSampler)
 *  dataset5: TotalThroughputSampler
 ```