feat: MaxAlloc improvements (#721)
Dependent on #719; can be reviewed before merging that.

## Which problem is this PR solving?

In v1, Refinery allowed setting a `MaxAlloc` value that controlled the
maximum amount of memory it would try to use. This is hard to keep
configured properly, because the guidance is to set it to some fraction
of the memory allocated to the pod. This PR implements two new values:

- the full allocation amount, called `AvailableMemory`
- the percentage refinery should use, called `MaxMemory`

The existing `MaxAlloc` value is still supported. If the first two are
nonzero, they are multiplied to get the maximum allocation; otherwise,
`MaxAlloc` is used. As in v1, if all of them are 0, refinery doesn't
attempt to limit its own use.

I spent a while trying to find a way to allow the app to read the max
allocation from the system, but it's pretty platform-dependent. It
seemed better to just make it so that the value can be put into the
environment in the Helm chart.

## Short description of the changes

- Add `AvailableMemory` and make it settable in the environment and on
the command line
- Implement the logic for calculating the value refinery should use and
use that in the allocation check
- Fix the names of the appropriate function calls and mocks
- Regenerate files